# rPackages: migrate code generation to use JSON