--- a/pkgs/development/r-modules/generate-shell.nix
+++ b/pkgs/development/r-modules/generate-shell.nix
@@ -17,6 +17,7 @@ stdenv.mkDerivation {
         data_table
         parallel
         BiocManager
+        jsonlite
       ];
     })
   ];