… pkgs/development/r-modules/bioc-annotation-packages.json
… pkgs/development/r-modules/bioc-annotation-packages.nix
… pkgs/development/r-modules/bioc-experiment-packages.json
… pkgs/development/r-modules/bioc-experiment-packages.nix
… pkgs/development/r-modules/bioc-packages.json
… pkgs/development/r-modules/bioc-packages.nix
… pkgs/development/r-modules/cran-packages.nix