## Description of changes

CC: @jbedo (current R codeowner)

This PR is an attempt to migrate all of the code gen of `rPackages` to use `.json`
This is in anticipation of the treewide reformat, which will affect all `.nix` files, including the huge ones for `rPackages`. By having it in `.json` we eliminate the need to worry about treewide formatting.

The first commit only migrates the generated code itself.
The second commit updates the docs
The third commit also migrates the generating code.

TODO:
- [x] use python to migrate to new json formatting
- [x] confirm that there are no rebuilds
- [x] benchmark performance impact on eval
- [x] update generating code

---
### Benchmarks
command: `hyperfine --runs 30 'nix-env -qaP -f . -A rPackages --no-name --out-path'`

This PR:
```
  Time (mean ± σ):     15.190 s ±  0.212 s    [User: 14.416 s, System: 0.712 s]
  Range (min … max):   14.980 s … 15.864 s    30 runs
```

Master:
```  
  Time (mean ± σ):     15.331 s ±  0.263 s    [User: 14.540 s, System: 0.736 s]
  Range (min … max):   14.908 s … 15.871 s    30 runs
```

---
 
The following script was used to generate the `.json` files from the `.nix` files. You'll need to change the file name for each package set. Also, this script doesn't yet generate the `extraArgs` and `packages` keys that have since been added.

<details>
<summary>The code</summary>

```
import re
import json

lines = open("bioc-experiment-packages.nix", "r").readlines()
lines=lines[8:-1]

def process_line(line):
    a = re.split(r';|,| |\[|\]|{|}|=|"', line.rstrip())
    a = [x for x in a if len(x) > 0]
    broken = a[-2] == "broken"
    if broken:
        a = a[:-2]
    nix_pname = a[0]

    di = {"name": a[3], "version": a[5], "sha256": a[7], "depends": a[9:]}

    x = [
        f'    "name": "{di["name"]}"',
        f'    "version": "{di["version"]}"',
        f'    "sha256": "{di["sha256"]}"',
        f'    "depends": {json.dumps(di["depends"])}',
    ]
    if broken:
        x.append(f'    "broken": true')
    y = ",\n".join(x)
    return f'  "{nix_pname}": {{\n{y}\n  }}'

print("{")
print(",\n".join(process_line(line) for line in lines))
print("}")

```
</details>

---

Here's my python code that can update the R packages:
https://gist.github.com/TomaSajt/6f49cc5c7c49606e2b712455a26a2621

<!--
For package updates please link to a changelog or describe changes, this helps your fellow maintainers discover breaking updates.
For new packages please briefly describe the package or provide a link to its homepage.
-->

## Things done

<!-- Please check what applies. Note that these are not hard requirements but merely serve as information for reviewers. -->

- Built on platform(s)
  - [ ] x86_64-linux
  - [ ] aarch64-linux
  - [ ] x86_64-darwin
  - [ ] aarch64-darwin
- For non-Linux: Is sandboxing enabled in `nix.conf`? (See [Nix manual](https://nixos.org/manual/nix/stable/command-ref/conf-file.html))
  - [ ] `sandbox = relaxed`
  - [ ] `sandbox = true`
- [ ] Tested, as applicable:
  - [NixOS test(s)](https://nixos.org/manual/nixos/unstable/index.html#sec-nixos-tests) (look inside [nixos/tests](https://github.com/NixOS/nixpkgs/blob/master/nixos/tests))
  - and/or [package tests](https://github.com/NixOS/nixpkgs/blob/master/pkgs/README.md#package-tests)
  - or, for functions and "core" functionality, tests in [lib/tests](https://github.com/NixOS/nixpkgs/blob/master/lib/tests) or [pkgs/test](https://github.com/NixOS/nixpkgs/blob/master/pkgs/test)
  - made sure NixOS tests are [linked](https://nixos.org/manual/nixpkgs/unstable/#ssec-nixos-tests-linking) to the relevant packages
- [ ] Tested compilation of all packages that depend on this change using `nix-shell -p nixpkgs-review --run "nixpkgs-review rev HEAD"`. Note: all changes have to be committed, also see [nixpkgs-review usage](https://github.com/Mic92/nixpkgs-review#usage)
- [ ] Tested basic functionality of all binary files (usually in `./result/bin/`)
- [24.11 Release Notes](https://github.com/NixOS/nixpkgs/blob/master/nixos/doc/manual/release-notes/rl-2411.section.md) (or backporting [23.11](https://github.com/NixOS/nixpkgs/blob/master/nixos/doc/manual/release-notes/rl-2311.section.md) and [24.05](https://github.com/NixOS/nixpkgs/blob/master/nixos/doc/manual/release-notes/rl-2405.section.md) Release notes)
  - [ ] (Package updates) Added a release notes entry if the change is major or breaking
  - [ ] (Module updates) Added a release notes entry if the change is significant
  - [ ] (Module addition) Added a release notes entry if adding a new NixOS module
- [ ] Fits [CONTRIBUTING.md](https://github.com/NixOS/nixpkgs/blob/master/CONTRIBUTING.md).

<!--
To help with the large amounts of pull requests, we would appreciate your
reviews of other pull requests, especially simple package updates. Just leave a
comment describing what you have tested in the relevant package/service.
Reviewing helps to reduce the average time-to-merge for everyone.
Thanks a lot if you do!

List of open PRs: https://github.com/NixOS/nixpkgs/pulls
Reviewing guidelines: https://github.com/NixOS/nixpkgs/blob/master/pkgs/README.md#reviewing-contributions
-->

---

Add a :+1: [reaction] to [pull requests you find important].

[reaction]: https://github.blog/2016-03-10-add-reactions-to-pull-requests-issues-and-comments/
[pull requests you find important]: https://github.com/NixOS/nixpkgs/pulls?q=is%3Aopen+sort%3Areactions-%2B1-desc