@jbedo 2024-07-18
> @jbedo what do you think, is this something that we should do?

This is a good change and a good opportunity to simplify the generation code by ditching the crude parser and using jsonlite. Happy to help with this.