@TomaSajt 2024-07-20
I pushed a commit that redoes the code-gen logic from scratch in python, since I was struggling with the R code.