@jbedo 2024-07-20
Sorry I don't think this is a good solution. Many members of the R community do not know python, and there's no technical reason not to continue to use R. Furthermore, this does not capture the logic of BIOC manager: BIOC releases are matched against specific R versions and it's important to use the correct release. Reimplementing details such as this in python would incur future maintenance burden. There are similar issues associated with reimplementing `availabe.packages()`.