@TomaSajt 2024-07-20
> Sorry I don't think this is a good solution. Many members of the R community do not know python, and there's no technical reason not to continue to use R. Furthermore, this does not capture the logic of BIOC manager: BIOC releases are matched against specific R versions and it's important to use the correct release. Reimplementing details such as this in python would incur future maintenance burden. There are similar issues associated with reimplementing `availabe.packages()`.

It's fine, since now I actually understand the logic behind the original script. This really was just a proof of concept of how we could structure the newer script.

By the way, I was having issues when trying out `jsonlite` because it always generated singleton lists instead of just the value I wanted... (e.g. the package name).
Maybe I just don't understand R enough and it's very simple.

Also, should we keep the current one line per package approach, or should we have one line per key. We no longer need a specific formatting to read the existing packages so it shouldn't be a problem either way.
However if we keep the one line per package approach the usual pretty and non-pretty formatting of json will not be what we want. We'll have to do something, like what's already being done and work line-by-line.