@jbedo 2024-07-22
I've updated the R version and pushed it to this branch.

> By the way, I was having issues when trying out `jsonlite` because it always generated singleton lists instead of just the value I wanted... (e.g. the package name). Maybe I just don't understand R enough and it's very simple.

It's annoying but singletons have to be `unbox()`ed.

> Also, should we keep the current one line per package approach, or should we have each key on a different line. We no longer need a specific formatting to read the existing packages so it shouldn't be a problem either way. However if we keep the one line per package approach the usual pretty and non-pretty formatting of json will not be what we want. We'll have to do something, like what's already being done and work line-by-line.

There's no reason to keep the one line format since we have good JSON parsers available. I think the current keyed by name format is fine and easy to work with.