@b-rodrigues 2024-07-22
I tried running the updater Rscript, and here is what I get:

```
{
  "1": {
    "depends": [
      "pbapply",
      "xtable"
    ],
    "name": "A3",
    "sha256": "017hq9pjsv1h9i7cqk5cfx27as54shlhdsdvr6jkhb8jfkpdb6cw",
    "version": "1.0.0"
  },
  "10": {
    "depends": [
      "kernlab"
    ],
    "name": "ABPS",
    "sha256": "0n3f66nmfi5v94il1mxy026mi84w01ph2aljk60vn3mrz8kwf2ll",
    "version": "0.3"
  },
  "100": {
    "depends": [
      "DISTRIB"
    ],
    "name": "AR",
    "sha256": "1wn4g2997c4vc3mzq2pv8ld0ryp0i78v4zlqqryvhshxgnz4f2pr",
    "version": "1.1"
  },
  "1000": {
    "depends": [
      "CEEMDANML"
    ],
    "name": "CompareMultipleModels",
    "sha256": "0qzybbxx5kqlrb260c0wh2vhp3ci5wk6853mm01wfyivhvf7sz0h",
    "version": "0.1.0"
  },
...
```

Just like with the old script, the names of the package aren't properly shown (they're replace by numbers here). I have no idea why. Only the broken packages are formatted correctly:

```
  "A3": {
    "broken": true,
    "depends": [
      "pbapply",
      "xtable"
    ],
    "name": "A3",
    "sha256": "017hq9pjsv1h9i7cqk5cfx27as54shlhdsdvr6jkhb8jfkpdb6cw",
    "version": "1.0.0"
  },
  "AATtools": {
    "broken": true,
    "depends": [
      "doParallel",
      "dplyr",
      "foreach",
      "magrittr"
    ],
    "name": "AATtools",
    "sha256": "10adwxyqnbjw2a0pxrp4qbs68fn2bdxwcbaqmvx5xfgrj4skl68z",
    "version": "0.0.2"
  },
  "ABACUS": {
    "broken": true,
    "depends": [
      "ggplot2",
      "shiny"
    ],
```

I have also reduced the number of workers, thinking it might be related to memory limits (I have 32gigs of ram though)