@TomaSajt 2024-07-22
I did some testing and compared the space requirements of the non-pretty and pretty versions.

I found that pretty printing will add around 1.6MB to the uncompressed repository size. (The r-modules directory went from 5376kB to 6868kB)

If we look at the compressed nixpkgs tarball's size, it increased by 52kB (42732kB to 42784kB).

This is probably still acceptable, just wanted to share it, in case it might be an issue.