@TomaSajt 2024-07-22
Okay, I just realized that the new R script puts everything into one long line.
I don't know if that's a good idea, since we can't really compare diffs like that.
Are you planning on changing that?