@b-rodrigues 2024-07-22
> Fixed, thanks.

This seems to work, but you need to remove line 81 in your script, which seems you added for testing purposes (I left a comment, but I don't know if you're properly notified)