@jbedo 2024-07-22
> That may be suboptimal because it will potentially cause spurious GIt conflicts. You probably want at least one line per package.

Good point, switched to pretty output.